ourVisDiff/requirements.txt
click
pandas
omegaconf
tqdm
wandb
lmdb
openai==0.28
Pillow
seaborn
scipy
scikit-learn
open_clip_torch
flask
salesforce-lavis